fix: add experimental helper from Flarum
Added an experimental helper from Flarum, which adds support for
setting the user one wants to run Composer as. This helper was
unfortunately deprecated when ynh_composer_as_user was officially added
to YunoHost, and the specific patch that added user support was not
migrated alongside with this. The patch is taken (and slightly
modified) from the following commit:

https://github.com/YunoHost-Apps/flarum_ynh/blob/0728ea096b4a0cd030b311821e659519b1346f4a/scripts/install